feat: Enhance OAuth integration with improved error handling and UI updates
